chore: 주석 추가 및 swagger관련 어노테이션 추가
coding standards에 맞게 주석 추가 및 swagger관련 어노테이션 추가

issue #3

+++ IrumOrder_BackEnd/src/main/java/Irumping/IrumOrder/repository/PayRepository.java
@@ -4,6 +4,5 @@ import Irumping.IrumOrder.entity.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PayRepository  extends JpaRepository<OrderEntity, Long> {
-    OrderEntity findByOrderIdAndUserId(int orderId, int userId); // 주문 ID와 사용자 ID로 주문 조회
-
+    OrderEntity findByOrderIdAndUserId(int orderId, long userId); // 주문 ID와 사용자 ID로 주문 조회
 }
